requirements.txt
ccxt
pandas
pyarrow
fastparquet
python-dotenv
mlflow
scikit-learn
xgboost
streamlit
sqlalchemy
psycopg2-binary
plotly
ipykernel
matplotlib
bayesian-optimization
tqdm
ta
seaborn
